Fix tip-thanks-modal showing behind top modal on iOS
