+++ app/app.tsx
@@ -45,7 +45,6 @@ import { useDeepLinking } from "./hooks/use-deep-linking"
 import { openActiveRide } from "./utils/helpers/ride-helpers"
 import { GestureHandlerRootView } from "react-native-gesture-handler"
 import { createModalStack, ModalProvider } from "react-native-modalfy"
-import { TipThanksModal } from "./screens/settings/components/tip-thanks-modal"
 import { RouteListWarningModal } from "./screens/route-list/components/route-list-warning-modal"
 import { DatePickerModal } from "./components/date-picker-modal/date-picker-modal.android"
 import { identifyPosthogUser, setAnalyticsUserProperty } from "./services/analytics"
@@ -54,7 +53,7 @@ enableScreens()
 
 export const queryClient = new QueryClient()
 
-const modalConfig = { TipThanksModal, RouteListWarningModal, DatePickerModal }
+const modalConfig = { RouteListWarningModal, DatePickerModal }
 const defaultOptions = { backdropOpacity: 0.6 }
 
 const stack = createModalStack(modalConfig, defaultOptions)
