+++ app/i18n/he.json
@@ -103,7 +103,7 @@
     "totalTips": "עד כה תרמתם",
     "thankYou": "תודה רבה!",
     "testflightMessage": "שימו לב: האפליקצייה הותקנה דרך TestFlight.\n על מנת לשלוח טיפ, יש להוריד את האפליקצייה דרך האפ סטור",
-    "thankYouText": "המון תודה על התרומה. זה מרגש וממלא במוטיבציה.‪\n\n שתהיה נסיעה טובה :-)\nגיא ומתן",
+    "thankYouText": "המון תודה על התרומה. זה מרגש וממלא במוטיבציה.‪\n\n שתהיה נסיעה טובה :-)\nגיא, מתן ודני",
     "aboutText": "בטר רייל היא אפליקציית קוד פתוח שנוצרה במטרה להנגיש את זמני רכבת ישראל לציבור",
     "independenceDeclaration": "האפליקצייה מפותחת באופן עצמאי ונכתבת בקוד פתוח",
     "follow": "עקבו אחרי Better Rail",
